+++ spec/test/components/dynamic/async/app_level_spec.rb
@@ -10,7 +10,6 @@ describe "Async Component", type: :feature, js: true do
         paragraph text: DateTime.now.strftime("%Q")
       end
     end
-
     expect(page).to have_content('Time now:')
     initial_timestamp = page.find("p").text # initial page load
     page.execute_script('MatestackUiCore.matestackEventHub.$emit("update")')
@@ -24,7 +23,6 @@ describe "Async Component", type: :feature, js: true do
         plain 'Time now: '
         paragraph text: DateTime.now.strftime("%Q")
       end
-      yield_page
     end
     matestack_render reset_app: false do
       plain 'A page inside the app'