+++ spec/test/components/dynamic/form/checkbox/custom_checkbox_spec.rb
@@ -21,15 +21,14 @@ describe "Form Component", type: :feature, js: true do
   before :each do
     allow_any_instance_of(FormTestController).to receive(:expect_params)
 
-    class Components::CustomFormCheckboxTest < Matestack::Ui::Core::Form::Checkbox::Base
-
-      vue_js_component_name "custom-form-checkbox-test"
+    class Components::CustomFormCheckboxTest < Matestack::Ui::VueJs::Components::Form::Checkbox
+      vue_name "custom-form-checkbox-test"
 
       def response
         div class: "custom-input-markup" do
-          label text: "my form input"
+          label "my form input"
           render_options
-          button attributes: {"@click": "changeValueViaJs(2)"}, text: "change value"
+          button "change value", "@click": "changeValueViaJs(2)", type: :button
           render_errors
         end
       end
@@ -44,16 +43,14 @@ describe "Form Component", type: :feature, js: true do
 
       class ExamplePage < Matestack::Ui::Page
         def response
-          form form_config do
+          m_form form_config do
             custom_form_checkbox_test key: :bar, id: "bar", options: { "Option 1": 1, "Option 2": 2 }
-            form_submit do
-              button text: 'Submit me!'
-            end
+            button 'Submit me!'
           end
         end
 
         def form_config
-          return {
+          {
             for: :my_object,
             method: :post,
             path: custom_checkbox_success_form_test_path
@@ -62,28 +59,21 @@ describe "Form Component", type: :feature, js: true do
       end
 
       visit '/example'
-
       expect(page).to have_xpath('//div[@class="custom-input-markup"]/input[@id="bar_1" and @class="js-added-class"]')
-
       click_button "change value"
-
       expect_any_instance_of(FormTestController).to receive(:expect_params)
         .with(hash_including(my_object: { bar: 2 }))
-
       click_button "Submit me!"
-
     end
 
     it "can be used within core form along other core inputs" do
 
       class ExamplePage < Matestack::Ui::Page
         def response
-          form form_config do
+          m_form form_config do
             form_checkbox key: :foo, id: "foo", options: { "Option 1": 1, "Option 2": 2 }
             custom_form_checkbox_test key: :bar, id: "bar", options: { "Option 3": 3, "Option 4": 4 }
-            form_submit do
-              button text: 'Submit me!'
-            end
+            button 'Submit me!'
           end
           toggle show_on: "form_submitted", id: 'async-form' do
             plain "form submitted!"
@@ -114,11 +104,9 @@ describe "Form Component", type: :feature, js: true do
     it "can display server errors async" do
       class ExamplePage < Matestack::Ui::Page
         def response
-          form form_config do
+          m_form form_config do
             custom_form_checkbox_test id: "foo", key: :foo, type: :text
-            form_submit do
-              button text: "Submit me!"
-            end
+            button "Submit me!"
           end
         end
 
@@ -134,18 +122,16 @@ describe "Form Component", type: :feature, js: true do
       visit "/example"
 
       click_button "Submit me!"
-      expect(page).to have_xpath('//span[@class="errors"]/span[@class="error" and contains(.,"seems to be invalid")]')
+      expect(page).to have_xpath('//div[@class="errors"]/div[@class="error" and contains(.,"seems to be invalid")]')
     end
 
     it "takes an array of options or hash and submits (multiple) selected item(s)" do
       class ExamplePage < Matestack::Ui::Page
         def response
-          form form_config, :include do
+          m_form form_config do
             custom_form_checkbox_test id: "my-array-test-checkbox", key: :array_input, options: ["Array Option 1","Array Option 2"]
             custom_form_checkbox_test id: "my-hash-test-checkbox", key: :hash_input, options: { "Hash Option 1": "1", "Hash Option 2": "2" }
-            form_submit do
-              button text: "Submit me!"
-            end
+            button "Submit me!"
           end
         end
 
@@ -171,12 +157,10 @@ describe "Form Component", type: :feature, js: true do
     it "can be initialized by (multiple) item(s)" do
       class ExamplePage < Matestack::Ui::Page
         def response
-            form form_config, :include do
+            m_form form_config do
               custom_form_checkbox_test id: "my-array-test-checkbox", key: :array_input, options: ["Array Option 1","Array Option 2"], init: ["Array Option 1", "Array Option 2"]
               custom_form_checkbox_test id: "my-hash-test-checkbox", key: :hash_input, options: { "Hash Option 1": "1", "Hash Option 2": "2" }, init: ["2"]
-              form_submit do
-                button text: "Submit me!"
-              end
+              button "Submit me!"
             end
         end
 
@@ -213,19 +197,14 @@ describe "Form Component", type: :feature, js: true do
       end
 
       class ExamplePage < Matestack::Ui::Page
-        def prepare
+        def response
           @test_model = TestModel.new
           @test_model.some_data = ["Array Option 2"]
           @test_model.more_data = ["my_second_key"]
-        end
-
-        def response
-          form form_config, :include do
+          m_form form_config do
             custom_form_checkbox_test id: "my-array-test-checkbox", key: :some_data, options: TestModel.array_options
             custom_form_checkbox_test id: "my-hash-test-checkbox", key: :more_data, options: TestModel.hash_options
-            form_submit do
-              button text: "Submit me!"
-            end
+            button "Submit me!"
           end
         end
 
@@ -246,13 +225,13 @@ describe "Form Component", type: :feature, js: true do
 
       uncheck "Hash Option 2"
       click_button "Submit me!"
-      expect(page).to have_xpath('//span[@class="errors"]/span[@class="error" and contains(.,"can\'t be blank")]')
+      expect(page).to have_xpath('//div[@class="errors"]/div[@class="error" and contains(.,"can\'t be blank")]')
 
       check "Hash Option 2"
       check "Hash Option 1"
       check "Array Option 1"
       click_button "Submit me!"
-      expect(page).not_to have_xpath('//span[@class="errors"]/span[@class="error" and contains(.,"can\'t be blank")]')
+      expect(page).not_to have_xpath('//div[@class="errors"]/div[@class="error" and contains(.,"can\'t be blank")]')
       #form should now be reset
       expect(page).to have_field('Array Option 1', checked: false)
       expect(page).to have_field('Array Option 2', checked: true)
@@ -267,16 +246,11 @@ describe "Form Component", type: :feature, js: true do
       load "#{Rails.root}/app/models/test_model.rb"
 
       class ExamplePage < Matestack::Ui::Page
-        def prepare
-          @test_model = TestModel.new
-        end
-
         def response
-          form form_config, :include do
+          @test_model = TestModel.new
+          m_form form_config do
             custom_form_checkbox_test id: "my-array-test-checkbox", key: :status, label: 'Status'
-            form_submit do
-              button text: "Submit me!"
-            end
+            button "Submit me!"
             toggle show_on: 'success', id: 'async-page' do
               plain 'Success'
             end