+++ CHANGELOG.md
@@ -23,7 +23,7 @@
 
 * `transition` can no longer handle symbols as path. Use rails path helper instead
 
-* `Matestack::Ui::DynamicActionviewComponent, Matestack::Ui::Core::Actionview::Dynamic` und static entfernt -> wird nicht benötigt, weil ActionView Helper zur Verfügung stehen
+* `Matestack::Ui::DynamicActionviewComponent, Matestack::Ui::Core::Actionview::Dynamic` and static removed -> is not needed
 
 * slots api has changed. Always use `method(:your_slot_method)` to pass slots into another component. Use `slot slots[:your_slot]` to render the slot. Pass optional params to slots in order to use with your method `slot slots[:your_slot], 'A param'`.